756 - Project 3 - Basics
Spring and JAX-WS annotations are added and most of the service's functionality is in place.
The layers are all created and organized.
Data Layer is complete.

Business Layer needs more work and the Service Layer needs a way to access incoming request headers for the authentication token hash.

SQLite Database needs to be created via schema.

This eclipse project is not yet a Dynamic Web Project yet. I.E. there is no web container yet which is required for the WAR creation.
Once the project is Web ready, spring configuration is needed for Bean creations.
For more help with that, visit: http://examples.javacodegeeks.com/enterprise-java/jws/jax-ws-spring-integration-example/