756 - Project 3 - Tweak Layers
Minor fixes in comments and code in the layers.
Re-generated java docs.

No longer attempting to acquire a SOAP header in the WebService.
Dean altered the project 3 design to incorporate the passing of the token hash in the service methods themselves for verification of authentication.

SQLite Database needs to be created via schema.

Still need the spring configuration files for deployment in the WAR.