756 - Project 3 - Admin Check
the setPrice method now checks for admin access for user.
Added the UserHasInsufficentPrivlegesException for when this fails.
(I'm expirmenting with this in the BeerService class, but I'm under the assumption if a method annotated with the @WebMethod annotation throws an Exception when being accessed, it will convert to a SOAP Fault. Though I'm not sure if that's true. If that's not the case, I will have to generate the Fault myself.)

Updated much of the JavaDocumentation.

Otherwise, same status as before:

*** SQLite database - Access the .db file from within the WAR file during deployment in GlassFish.
*** Properties File - Get these values into code.
*** Spring Beans    - Is this still worth pursuing?