756 - Project 3 - SOAP Faults
Made all methods in the BeerService throw exceptions if instead of handle them inside of try/catch blocks.
In this way, the generated WSDL will know to return SOAP Fault responses when a client causes one of these methods to throw an exception.
By default, the SOAP Fault's message body is the thrown exception's message.