756 - Project 3 - Client Built & Server Tested
For as many cases as I could think of, this server is tested.
I caved and ended up using JodaTime (http://www.joda.org/joda-time/) for a portion of the code.
(JDK 7 has poor time manipulation support.)

I used Eclipse to Clean Up the source files. So they are all modified.
I did add one more exception (MismatchingPasswordException) and re-assessed where certain code should have resided.
Particularly, I moved the age checking from the Data Layer to the getToken method in the Business Layer where it should be.

The documentation is tidied up.
The WAR is built and deployable.

Included is also a sub-project containing the client for accessing the BeerService.
The index.php file contains explicit instructions on how to deploy the WAR (correctly) and how to utilize my client somewhat.